GHA: Add IRC notifications action
This uses a composite action with conditionals, but these will not yet
work per <https://github.com/actions/runner/issues/834>.

This commit is just to show the start of the action before transforming
it using a shell script workaround. If simpler, that transformation can
be reverted to this version when support for conditionals in composite
actions is added.